@@ -113,38 +113,38 @@
     "@angular/cli": "^13.3.4",
     "@angular/compiler-cli": "^13.3.5",
     "@angular/language-service": "^13.3.5",
-    "@nrwl/angular": "14.0.5",
-    "@nrwl/cli": "14.0.5",
-    "@nrwl/cypress": "14.0.5",
-    "@nrwl/eslint-plugin-nx": "14.0.5",
-    "@nrwl/jest": "14.0.5",
-    "@nrwl/linter": "14.0.5",
-    "@nrwl/nx-cloud": "14.0.1",
-    "@nrwl/workspace": "14.0.5",
-    "@types/jest": "27.4.1",
-    "@types/node": "^16.11.32",
-    "@typescript-eslint/eslint-plugin": "^5.21.0",
-    "@typescript-eslint/parser": "^5.21.0",
+    "@nrwl/angular": "14.1.9",
+    "@nrwl/cli": "14.1.9",
+    "@nrwl/cypress": "14.1.9",
+    "@nrwl/eslint-plugin-nx": "14.1.9",
+    "@nrwl/jest": "14.1.9",
+    "@nrwl/linter": "14.1.9",
+    "@nrwl/nx-cloud": "14.0.5",
+    "@nrwl/workspace": "14.1.9",
+    "@types/jest": "27.5",
+    "@types/node": "^16.11.36",
+    "@typescript-eslint/eslint-plugin": "^5.26.0",
+    "@typescript-eslint/parser": "^5.26.0",
     "cypress": "^9.6.0",
-    "eslint": "8.14",
+    "eslint": "8.16",
     "eslint-config-prettier": "8.5",
     "eslint-plugin-cypress": "^2.10.3",
     "eslint-plugin-ngrx": "^2.1.2",
     "imng-schematics": "^4.118.3",
-    "jest": "27.5.1",
+    "jest": "^27.5.1",
     "jest-junit": "^13.2.0",
     "jest-preset-angular": "^11.1.2",
     "ng-packagr": "13.3.1",
     "ngx-bootstrap": "^8.0.0",
-    "nx": "14.0.5",
-    "openapi-mermaid": "^2.115.3",
-    "openapi-ts-generator": "^4.114.2",
-    "postcss": "^8.4.12",
-    "postcss-import": "^14.1.0",
-    "postcss-preset-env": "^7.4.4",
-    "postcss-url": "^10.1.1",
+    "nx": "14.1.9",
+    "openapi-mermaid": "^2.145.7",
+    "openapi-ts-generator": "^4.146.2",
+    "postcss": "^8.4.14",
+    "postcss-import": "14.1.0",
+    "postcss-preset-env": "7.6",
+    "postcss-url": "10.1.3",
     "prettier": "^2.6.2",
-    "ts-jest": "27.1.4",
+    "ts-jest": "^27.1.5",
     "typescript": "4.6.3"
   },
   "dependencies": {
@@ -159,31 +159,31 @@
     "@angular/router": "^13.3.5",
     "@angular/service-worker": "^13.3.5",
     "@ericblade/quagga2": "^1.7.0",
-    "@microsoft/applicationinsights-web": "^2.8.1",
-    "@microsoft/signalr": "^6.0.4",
+    "@microsoft/applicationinsights-web": "^2.8.3",
+    "@microsoft/signalr": "^6.0.5",
     "@ngrx/effects": "^13.1.0",
     "@ngrx/router-store": "^13.1.0",
     "@ngrx/store": "^13.1.0",
     "@ngrx/store-devtools": "^13.1.0",
     "@progress/kendo-angular-buttons": "^8.0.0",
     "@progress/kendo-angular-charts": "^7.1.0",
-    "@progress/kendo-angular-common": "^3.0.0",
-    "@progress/kendo-angular-dateinputs": "^7.0.0",
+    "@progress/kendo-angular-common": "^3.1.0",
+    "@progress/kendo-angular-dateinputs": "^7.0.1",
     "@progress/kendo-angular-dialog": "^7.0.0",
-    "@progress/kendo-angular-dropdowns": "^7.0.0",
+    "@progress/kendo-angular-dropdowns": "^7.0.1",
     "@progress/kendo-angular-excel-export": "^5.0.0",
-    "@progress/kendo-angular-grid": "^7.0.0",
-    "@progress/kendo-angular-inputs": "^9.0.0",
+    "@progress/kendo-angular-grid": "^7.0.1",
+    "@progress/kendo-angular-inputs": "^9.0.1",
     "@progress/kendo-angular-intl": "^4.0.0",
     "@progress/kendo-angular-l10n": "^4.0.0",
     "@progress/kendo-angular-label": "^4.0.0",
     "@progress/kendo-angular-pdf-export": "^4.0.0",
     "@progress/kendo-angular-popup": "^5.0.0",
-    "@progress/kendo-angular-treeview": "^7.0.0",
+    "@progress/kendo-angular-treeview": "^7.0.1",
     "@progress/kendo-data-query": "^1.5.6",
     "@progress/kendo-drawing": "^1.16.3",
     "@progress/kendo-licensing": "^1.2.2",
-    "@progress/kendo-theme-default": "^5.3.1",
+    "@progress/kendo-theme-default": "^5.4.1",
     "ang-jsoneditor": "^1.10.5",
     "jsoneditor": "^9.7.4",
     "ngx-barcode-scanner": "^0.3.0",
