@@ -1,4 +1,4 @@
-module.exports = {
+export default {
   displayName: 'imng-kendo-chart-odata',
 
   reporters: [
@@ -29,5 +29,5 @@ module.exports = {
     'jest-preset-angular/build/serializers/ng-snapshot',
     'jest-preset-angular/build/serializers/html-comment',
   ],
-  preset: '../../jest.preset.ts',
+  preset: '../../jest.preset.js',
 };
