@@ -26,16 +26,13 @@ export function OrderFormGroupFac(): FormGroup {
     customerId: new FormControl('', Validators.required),
     shipToAddressId: new FormControl(''),
     billToAddressId: new FormControl(''),
-    shipMethod: new FormControl(
-      '',
-      Validators.compose([Validators.required, Validators.maxLength(50)]),
-    ),
     creditCardApprovalCode: new FormControl('', Validators.maxLength(15)),
     subTotal: new FormControl('', Validators.required),
     taxAmt: new FormControl('', Validators.required),
     freight: new FormControl('', Validators.required),
     totalDue: new FormControl('', Validators.required),
     comment: new FormControl('', Validators.maxLength(500)),
+    shippingType: new FormControl('', Validators.required),
     customer: new FormControl(''),
     shipToAddress: new FormControl(''),
     billToAddress: new FormControl(''),
