@@ -20,13 +20,13 @@ export enum OrderProperties {
   CUSTOMER_ID = 'customerId',
   SHIP_TO_ADDRESS_ID = 'shipToAddressId',
   BILL_TO_ADDRESS_ID = 'billToAddressId',
-  SHIP_METHOD = 'shipMethod',
   CREDIT_CARD_APPROVAL_CODE = 'creditCardApprovalCode',
   SUB_TOTAL = 'subTotal',
   TAX_AMT = 'taxAmt',
   FREIGHT = 'freight',
   TOTAL_DUE = 'totalDue',
   COMMENT = 'comment',
+  SHIPPING_TYPE = 'shippingType',
   CUSTOMER = 'customer',
   SHIP_TO_ADDRESS = 'shipToAddress',
   BILL_TO_ADDRESS = 'billToAddress',
