@@ -5,9 +5,10 @@
  * For issues or feature request, visit the repo: https://github.com/ikemtz/openapi-ts-generator
  * Do not edit.
  */
-// eslint-disable-next-line @typescript-eslint/no-empty-interface
+
 import { ICustomer } from './customer.model';
 
+// eslint-disable-next-line @typescript-eslint/no-empty-interface
 export interface ISalesAgent {
   id?: number;
   name?: string;
