@@ -8,7 +8,7 @@ async function generateTsModels() {
     openApiJsonUrl:
       'https://im-wa-crto-nrcrn.azurewebsites.net/swagger/v1/swagger.json',
     outputPath: './apps/nurse-cron/src/app/models/certifications-odata/',
-    typeFilterCallBack: (val, i, arr) => !val.name.endsWith('ODataEnvelope'),
+    typeFilterCallBack: tsGenerator.nrsrxTypeFilterCallBack,
     valuePropertyTypeFilterCallBack: (val, i, arr) =>
       !val.name.startsWith('created') && !val.name.startsWith('updated'),
     genAngularFormGroups: true,
@@ -19,7 +19,7 @@ async function generateTsModels() {
     openApiJsonUrl:
       'https://im-wa-crto-nrcrn.azurewebsites.net/swagger/v1/swagger.json',
     outputPath: './apps/nurse-cron/diagrams/certifications-odata/',
-    typeFilterCallBack: (val, i, arr) => !val.name.endsWith('ODataEnvelope'),
+    typeFilterCallBack: tsGenerator.nrsrxTypeFilterCallBack,
   });
 
   //Competencies
@@ -28,7 +28,7 @@ async function generateTsModels() {
     openApiJsonUrl:
       'https://im-wa-cmpo-nrcrn.azurewebsites.net/swagger/v1/swagger.json',
     outputPath: './apps/nurse-cron/src/app/models/competencies-odata/',
-    typeFilterCallBack: (val, i, arr) => !val.name.endsWith('ODataEnvelope'),
+    typeFilterCallBack: tsGenerator.nrsrxTypeFilterCallBack,
     valuePropertyTypeFilterCallBack: (val, i, arr) =>
       !val.name.startsWith('created') && !val.name.startsWith('updated'),
     genAngularFormGroups: true,
@@ -39,7 +39,7 @@ async function generateTsModels() {
     openApiJsonUrl:
       'https://im-wa-crto-nrcrn.azurewebsites.net/swagger/v1/swagger.json',
     outputPath: './apps/nurse-cron/diagrams/competencies-odata/',
-    typeFilterCallBack: (val, i, arr) => !val.name.endsWith('ODataEnvelope'),
+    typeFilterCallBack: tsGenerator.nrsrxTypeFilterCallBack,
   });
 
   //Employees
@@ -48,7 +48,7 @@ async function generateTsModels() {
     openApiJsonUrl:
       'https://im-wa-empo-nrcrn.azurewebsites.net/swagger/v1/swagger.json',
     outputPath: './apps/nurse-cron/src/app/models/employees-odata/',
-    typeFilterCallBack: (val, i, arr) => !val.name.endsWith('ODataEnvelope'),
+    typeFilterCallBack: tsGenerator.nrsrxTypeFilterCallBack,
     valuePropertyTypeFilterCallBack: (val, i, arr) =>
       !val.name.startsWith('created') && !val.name.startsWith('updated'),
     genAngularFormGroups: true,
@@ -59,7 +59,7 @@ async function generateTsModels() {
     openApiJsonUrl:
       'https://im-wa-empo-nrcrn.azurewebsites.net/swagger/v1/swagger.json',
     outputPath: './apps/nurse-cron/diagrams/employees-odata/',
-    typeFilterCallBack: (val, i, arr) => !val.name.endsWith('ODataEnvelope'),
+    typeFilterCallBack: tsGenerator.nrsrxTypeFilterCallBack,
   });
 
   //HealthItems
@@ -68,7 +68,7 @@ async function generateTsModels() {
     openApiJsonUrl:
       'https://im-wa-hlto-nrcrn.azurewebsites.net/swagger/v1/swagger.json',
     outputPath: './apps/nurse-cron/src/app/models/health-items-odata/',
-    typeFilterCallBack: (val, i, arr) => !val.name.endsWith('ODataEnvelope'),
+    typeFilterCallBack: tsGenerator.nrsrxTypeFilterCallBack,
     valuePropertyTypeFilterCallBack: (val, i, arr) =>
       !val.name.startsWith('created') && !val.name.startsWith('updated'),
     genAngularFormGroups: true,
@@ -79,7 +79,7 @@ async function generateTsModels() {
     openApiJsonUrl:
       'https://im-wa-hlto-nrcrn.azurewebsites.net/swagger/v1/swagger.json',
     outputPath: './apps/nurse-cron/diagrams/health-items-odata/',
-    typeFilterCallBack: (val, i, arr) => !val.name.endsWith('ODataEnvelope'),
+    typeFilterCallBack: tsGenerator.nrsrxTypeFilterCallBack,
   });
 
   //Schedules
@@ -88,7 +88,7 @@ async function generateTsModels() {
     openApiJsonUrl:
       'https://im-wa-scdo-nrcrn.azurewebsites.net/swagger/v1/swagger.json',
     outputPath: './apps/nurse-cron/src/app/models/schedules-odata/',
-    typeFilterCallBack: (val, i, arr) => !val.name.endsWith('ODataEnvelope'),
+    typeFilterCallBack: tsGenerator.nrsrxTypeFilterCallBack,
     valuePropertyTypeFilterCallBack: (val, i, arr) =>
       !val.name.startsWith('created') && !val.name.startsWith('updated'),
     genAngularFormGroups: true,
@@ -99,7 +99,7 @@ async function generateTsModels() {
     openApiJsonUrl:
       'https://im-wa-scdo-nrcrn.azurewebsites.net/swagger/v1/swagger.json',
     outputPath: './apps/nurse-cron/diagrams/schedules-odata/',
-    typeFilterCallBack: (val, i, arr) => !val.name.endsWith('ODataEnvelope'),
+    typeFilterCallBack: tsGenerator.nrsrxTypeFilterCallBack,
   });
 
   //Units
@@ -108,7 +108,7 @@ async function generateTsModels() {
     openApiJsonUrl:
       'https://im-wa-unto-nrcrn.azurewebsites.net/swagger/v1/swagger.json',
     outputPath: './apps/nurse-cron/src/app/models/units-odata/',
-    typeFilterCallBack: (val, i, arr) => !val.name.endsWith('ODataEnvelope'),
+    typeFilterCallBack: tsGenerator.nrsrxTypeFilterCallBack,
     valuePropertyTypeFilterCallBack: (val, i, arr) =>
       !val.name.startsWith('created') && !val.name.startsWith('updated'),
     genAngularFormGroups: true,
@@ -119,7 +119,7 @@ async function generateTsModels() {
     openApiJsonUrl:
       'https://im-wa-unto-nrcrn.azurewebsites.net/swagger/v1/swagger.json',
     outputPath: './apps/nurse-cron/diagrams/units-odata/',
-    typeFilterCallBack: (val, i, arr) => !val.name.endsWith('ODataEnvelope'),
+    typeFilterCallBack: tsGenerator.nrsrxTypeFilterCallBack,
   });
 }
 
