@@ -1,3 +1,4 @@
+/* istanbul ignore file */
 /**
  * This file is generated by the openapi-ts-generator
  * #form-group-factory.ts.hbs
@@ -10,13 +11,22 @@ import { FormControl, FormArray, FormGroup, Validators } from '@angular/forms'; 
 export function EmployeeFormGroupFac(): FormGroup {
   return new FormGroup({
     id: new FormControl(''),
-    lastName: new FormControl('', Validators.compose([Validators.required, Validators.maxLength(250)])),
-    firstName: new FormControl('', Validators.compose([Validators.required, Validators.maxLength(250)])),
+    lastName: new FormControl(
+      '',
+      Validators.compose([Validators.required, Validators.maxLength(250)]),
+    ),
+    firstName: new FormControl(
+      '',
+      Validators.compose([Validators.required, Validators.maxLength(250)]),
+    ),
     birthDate: new FormControl(''),
     mobilePhone: new FormControl('', Validators.maxLength(25)),
     homePhone: new FormControl('', Validators.maxLength(25)),
     photo: new FormControl('', Validators.maxLength(4000)),
-    email: new FormControl('', Validators.compose([Validators.required, Validators.maxLength(250)])),
+    email: new FormControl(
+      '',
+      Validators.compose([Validators.required, Validators.maxLength(250)]),
+    ),
     addressLine1: new FormControl('', Validators.maxLength(250)),
     addressLine2: new FormControl('', Validators.maxLength(250)),
     city: new FormControl('', Validators.maxLength(150)),
