@@ -1,14 +1,16 @@
+/* istanbul ignore file */
 /**
  * This file is generated by the openapi-ts-generator
  * #model.ts.hbs
  * For issues or feature request, visit the repo: https://github.com/ikemtz/openapi-ts-generator
  * Do not edit.
  */
-// eslint-disable-next-line @typescript-eslint/no-empty-interface
+
 import { IEmployeeCertification } from './employee-certification.model';
 import { IEmployeeCompetency } from './employee-competency.model';
 import { IEmployeeHealthItem } from './employee-health-item.model';
 
+// eslint-disable-next-line @typescript-eslint/no-empty-interface
 export interface IEmployee {
   id?: string;
   lastName?: string;
