@@ -1,3 +1,4 @@
+/* istanbul ignore file */
 /**
  * This file is generated by the openapi-ts-generator
  * #form-group-factory.ts.hbs
@@ -11,9 +12,15 @@ export function ScheduleFormGroupFac(): FormGroup {
   return new FormGroup({
     id: new FormControl(''),
     unitId: new FormControl('', Validators.required),
-    unitName: new FormControl('', Validators.compose([Validators.required, Validators.maxLength(250)])),
+    unitName: new FormControl(
+      '',
+      Validators.compose([Validators.required, Validators.maxLength(250)]),
+    ),
     employeeId: new FormControl('', Validators.required),
-    employeeName: new FormControl('', Validators.compose([Validators.required, Validators.maxLength(250)])),
+    employeeName: new FormControl(
+      '',
+      Validators.compose([Validators.required, Validators.maxLength(250)]),
+    ),
     staffingRequirementId: new FormControl('', Validators.required),
     startTimeUtc: new FormControl('', Validators.required),
     scheduledHours: new FormControl('', Validators.required),
