@@ -1,3 +1,4 @@
+/* istanbul ignore file */
 /**
  * This file is generated by the openapi-ts-generator
  * #form-group-factory.ts.hbs
@@ -10,14 +11,35 @@ import { FormControl, FormArray, FormGroup, Validators } from '@angular/forms'; 
 export function BuildingFormGroupFac(): FormGroup {
   return new FormGroup({
     id: new FormControl(''),
-    name: new FormControl('', Validators.compose([Validators.required, Validators.maxLength(50)])),
+    name: new FormControl(
+      '',
+      Validators.compose([Validators.required, Validators.maxLength(50)]),
+    ),
     siteName: new FormControl('', Validators.maxLength(150)),
-    addressLine1: new FormControl('', Validators.compose([Validators.required, Validators.maxLength(250)])),
-    addressLine2: new FormControl('', Validators.compose([Validators.required, Validators.maxLength(250)])),
-    cityOrMunicipality: new FormControl('', Validators.compose([Validators.required, Validators.maxLength(250)])),
-    stateOrProvidence: new FormControl('', Validators.compose([Validators.required, Validators.maxLength(50)])),
-    postalCode: new FormControl('', Validators.compose([Validators.required, Validators.maxLength(50)])),
-    country: new FormControl('', Validators.compose([Validators.required, Validators.maxLength(3)])),
+    addressLine1: new FormControl(
+      '',
+      Validators.compose([Validators.required, Validators.maxLength(250)]),
+    ),
+    addressLine2: new FormControl(
+      '',
+      Validators.compose([Validators.required, Validators.maxLength(250)]),
+    ),
+    cityOrMunicipality: new FormControl(
+      '',
+      Validators.compose([Validators.required, Validators.maxLength(250)]),
+    ),
+    stateOrProvidence: new FormControl(
+      '',
+      Validators.compose([Validators.required, Validators.maxLength(50)]),
+    ),
+    postalCode: new FormControl(
+      '',
+      Validators.compose([Validators.required, Validators.maxLength(50)]),
+    ),
+    country: new FormControl(
+      '',
+      Validators.compose([Validators.required, Validators.maxLength(3)]),
+    ),
     gpsData: new FormControl(''),
     deletedBy: new FormControl('', Validators.maxLength(250)),
     deletedOnUtc: new FormControl(''),
