@@ -1,3 +1,4 @@
+/* istanbul ignore file */
 /**
  * This file is generated by the openapi-ts-generator
  * #form-group-factory.ts.hbs
@@ -11,7 +12,10 @@ export function UnitFormGroupFac(): FormGroup {
   return new FormGroup({
     id: new FormControl(''),
     buildingId: new FormControl('', Validators.required),
-    name: new FormControl('', Validators.compose([Validators.required, Validators.maxLength(102)])),
+    name: new FormControl(
+      '',
+      Validators.compose([Validators.required, Validators.maxLength(102)]),
+    ),
     roomCount: new FormControl('', Validators.required),
     deletedBy: new FormControl('', Validators.maxLength(250)),
     deletedOnUtc: new FormControl(''),
