@@ -1,5 +1,5 @@
 export * from './lib/imng-oidc-client.module';
-export * from './lib/+state/oidc-user.facade';
+export { OidcUserFacade } from './lib/+state/oidc-user.facade';
 export { OidcUserSelectors } from './lib/+state/oidc-user.selectors';
 export * as oidcActionTypes from './lib/+state/oidc.actions';
 export * from './lib/+state/oidc.effects';
